- use QTDIR environment for build process

+++ sax-head/spp/Makefile.lex
@@ -10,24 +10,25 @@ swig:spp.o spp.tab.o lex.yy.o
 
 spp:spp.tab.o lex.yy.o spp.o example.o
 	g++ -Wall -O2 -o example example.o spp.o spp.tab.o lex.yy.o \
-		-L/usr/lib/qt3/lib -I/usr/lib/qt3/include -lfl -lreadline -lqt-mt
+		-L${QTDIR}/lib -I${QTDIR}/include -lfl -lreadline -lqt-mt
 
 example.o:example.cpp
-	g++ -Wall -O2 -fpic -c example.cpp -I. -I/usr/lib/qt3/include
+	g++ -Wall -O2 -fpic -c example.cpp -I. -I${QTDIR}/include
 
 spp.o:spp.cpp spp.h
-	g++ -Wall -O2 -fpic -c spp.cpp -I. -I/usr/lib/qt3/include
+	g++ -Wall -O2 -fpic -c spp.cpp -I. -I${QTDIR}/include
 
 spp.tab.o: spp.tab.cc lex.yy.c
-	g++ -Wall -O2 -fpic -c spp.tab.cc -I. -I/usr/lib/qt3/include
+	g++ -Wall -O2 -fpic -c spp.tab.cc -I. -I${QTDIR}/include
 
 lex.yy.o: lex.yy.c
-	g++ -Wall -O2 -fpic -c lex.yy.c -I. -I/usr/lib/qt3/include
+	g++ -Wall -O2 -fpic -c lex.yy.c -I. -I${QTDIR}/include
 
 spp.tab.cc: spp.yc
 	bison -d spp.yc
+	grep -qi "goto yyerrlab1;" spp.tab.cc || \
 	cat spp.tab.cc | sed -e s@yyerrlab1:@@ > spp.tab.cc.new
-	mv spp.tab.cc.new spp.tab.cc
+	test -f spp.tab.cc.new && mv spp.tab.cc.new spp.tab.cc || true
 
 lex.yy.c: spp.l spp.tab.cc
 	flex -i spp.l
